requirements.txt
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.8
python-dotenv==1.0.0
pytest==7.3.1
pytest-mock==3.10.0
backtrader==1.9.78.123
pyotp==2.8.0
websocket-client==1.5.1
requests==2.31.0
pymongo==4.3.3
loguru==0.7.0
tqdm==4.65.0
python-dateutil==2.8.2
joblib==1.2.0
scikit-learn==1.2.2
